fix importlib error and use dict
